[refactory] Renomeando pacote de integrações externas
