--- src/main/java/br/com/nivlabs/gp/client/RestClient.java
@@ -1,5 +1,0 @@
-package br.com.nivlabs.gp.client;
-
-public interface RestClient {
-
-}
