+++ src/main/java/br/com/nivlabs/gp/integration/RestClient.java
@@ -1,0 +1,5 @@
+package br.com.nivlabs.gp.integration;
+
+public interface RestClient {
+
+}
